+++ lib/src/network_util.dart
@@ -1,10 +1,9 @@
 import 'dart:convert';
 
+import 'package:flutter_polyline_points/src/utils/polyline_decoder.dart';
+import 'package:flutter_polyline_points/src/utils/polyline_request.dart';
 import 'package:http/http.dart' as http;
 
-import '../src/PointLatLng.dart';
-import '../src/utils/polyline_waypoint.dart';
-import '../src/utils/request_enums.dart';
 import 'utils/polyline_result.dart';
 
 class NetworkUtil {
@@ -12,88 +11,55 @@ class NetworkUtil {
 
   ///Get the encoded string from google directions api
   ///
-  Future<PolylineResult> getRouteBetweenCoordinates(
-      String googleApiKey,
-      PointLatLng origin,
-      PointLatLng destination,
-      TravelMode travelMode,
-      List<PolylineWayPoint> wayPoints,
-      bool avoidHighways,
-      bool avoidTolls,
-      bool avoidFerries,
-      bool optimizeWaypoints) async {
-    String mode = travelMode.toString().replaceAll('TravelMode.', '');
-    PolylineResult result = PolylineResult();
-    var params = {
-      "origin": "${origin.latitude},${origin.longitude}",
-      "destination": "${destination.latitude},${destination.longitude}",
-      "mode": mode,
-      "avoidHighways": "$avoidHighways",
-      "avoidFerries": "$avoidFerries",
-      "avoidTolls": "$avoidTolls",
-      "key": googleApiKey
-    };
-    if (wayPoints.isNotEmpty) {
-      List wayPointsArray = [];
-      wayPoints.forEach((point) => wayPointsArray.add(point.location));
-      String wayPointsString = wayPointsArray.join('|');
-      if (optimizeWaypoints) {
-        wayPointsString = 'optimize:true|$wayPointsString';
-      }
-      params.addAll({"waypoints": wayPointsString});
-    }
-    Uri uri =
-        Uri.https("maps.googleapis.com", "maps/api/directions/json", params);
+  Future<List<PolylineResult>> getRouteBetweenCoordinates({
+    required PolylineRequest request,
+    String? googleApiKey,
+  }) async {
+    List<PolylineResult> results = [];
 
-    // print('GOOGLE MAPS URL: ' + url);
-    var response = await http.get(uri);
+    var response = await http.get(
+      request.toUri(apiKey: googleApiKey),
+      headers: request.headers,
+    );
     if (response.statusCode == 200) {
       var parsedJson = json.decode(response.body);
-      result.status = parsedJson["status"];
       if (parsedJson["status"]?.toLowerCase() == STATUS_OK &&
           parsedJson["routes"] != null &&
           parsedJson["routes"].isNotEmpty) {
-        result.points = decodeEncodedPolyline(
-            parsedJson["routes"][0]["overview_polyline"]["points"]);
+        List<dynamic> routeList = parsedJson["routes"];
+        for (var route in routeList) {
+          results.add(PolylineResult(
+            points: PolylineDecoder.run(route["overview_polyline"]["points"]),
+            errorMessage: "",
+            status: parsedJson["status"],
+            totalDistanceValue: route['legs']
+                .map((leg) => leg['distance']['value'])
+                .reduce((v1, v2) => v1 + v2),
+            distanceTexts: <String>[
+              ...route['legs'].map((leg) => leg['distance']['text'])
+            ],
+            distanceValues: <int>[
+              ...route['legs'].map((leg) => leg['distance']['value'])
+            ],
+            overviewPolyline: route["overview_polyline"]["points"],
+            totalDurationValue: route['legs']
+                .map((leg) => leg['duration']['value'])
+                .reduce((v1, v2) => v1 + v2),
+            durationTexts: <String>[
+              ...route['legs'].map((leg) => leg['duration']['text'])
+            ],
+            durationValues: <int>[
+              ...route['legs'].map((leg) => leg['duration']['value'])
+            ],
+            endAddress: route["legs"].last['end_address'],
+            startAddress: route["legs"].first['start_address'],
+          ));
+        }
       } else {
-        result.errorMessage = parsedJson["error_message"];
+        throw Exception(
+            "Unable to get route: Response ---> ${parsedJson["status"]} ");
       }
     }
-    return result;
-  }
-
-  ///decode the google encoded string using Encoded Polyline Algorithm Format
-  /// for more info about the algorithm check https://developers.google.com/maps/documentation/utilities/polylinealgorithm
-  ///
-  ///return [List]
-  List<PointLatLng> decodeEncodedPolyline(String encoded) {
-    List<PointLatLng> poly = [];
-    int index = 0, len = encoded.length;
-    int lat = 0, lng = 0;
-
-    while (index < len) {
-      int b, shift = 0, result = 0;
-      do {
-        b = encoded.codeUnitAt(index++) - 63;
-        result |= (b & 0x1f) << shift;
-        shift += 5;
-      } while (b >= 0x20);
-      int dlat = ((result & 1) != 0 ? ~(result >> 1) : (result >> 1));
-      lat += dlat;
-
-      shift = 0;
-      result = 0;
-      do {
-        b = encoded.codeUnitAt(index++) - 63;
-        result |= (b & 0x1f) << shift;
-        shift += 5;
-      } while (b >= 0x20);
-      int dlng = ((result & 1) != 0 ? ~(result >> 1) : (result >> 1));
-      lng += dlng;
-      PointLatLng p =
-          new PointLatLng((lat / 1E5).toDouble(), (lng / 1E5).toDouble());
-      poly.add(p);
-    }
-    return poly;
+    return results;
   }
 }